+++ Servers/services/reportService.ts
@@ -3,9 +3,10 @@ import { DefaultReportName, ReportType } from "../models/reporting.model";
 import { getVendorRisksByProjectIdQuery } from "../utils/vendorRisk.util";
 import { getVendorByProjectIdQuery } from "../utils/vendor.utils";
 import { getUserByIdQuery } from "../utils/user.utils";
-import {getAssessmentTrackerMarkdown} from './markdowns/assessmentTrackerMarkdown';
+import { getProjectRiskMarkdown } from "./markdowns/projectRiskMarkdown";
+import { getAssessmentTrackerMarkdown } from './markdowns/assessmentTrackerMarkdown';
 
-interface reportBodyData {
+export interface reportBodyData {
   projectTitle: string,
   projectOwner: string
 }
@@ -87,7 +88,7 @@ export async function getReportData(
   let markdownFormattedData;
   switch(reportType) {
     case ReportType.PROJECTRISK_REPORT:
-      markdownFormattedData = getProjectRiskMarkdown(projectId, reportBody)
+      markdownFormattedData = await getProjectRiskMarkdown(projectId, reportBody)
       break;
     case ReportType.ASSESSMENT_REPORT:
       markdownFormattedData = await getAssessmentTrackerMarkdown(projectId, frameworkId, reportBody)
@@ -101,42 +102,7 @@ export async function getReportData(
   return markdownFormattedData;
 }
   
-/*
-  This is the markdown data for project risk report
-*/
-export async function getProjectRiskMarkdown (
-    projectId: number,
-    data: reportBodyData
-  ) : Promise<any> {
-  let rows: string;
-  const reportData = await getProjectRisksReportQuery(projectId);
 
-  if (reportData.length > 0) {
-    rows = reportData.map((risk: { risk_name: any; risk_owner: any; risk_severity: any; likelihood: any; approval_status: any; risk_level_autocalculated: any; deadline: { toLocaleDateString: () => any; }; }) => 
-      `| ${risk.risk_name} | ${risk.risk_owner} | ${risk.risk_severity} | ${risk.likelihood} | ${risk.approval_status} | ${risk.risk_level_autocalculated} | ${risk.deadline.toLocaleDateString()} |`
-    ).join('\n');
-  } else {
-    rows = `| - | - | - | - | - | - | - |`
-  }
-
-  const projectRiskMD = `
-VerifyWise project risk report
-========================
-
-This report is generated by the VeriftyWise Project Risk. It aims to provide a way to demonstrate their claims about the risks of their AI systems.
-  
-- **Date** ${new Date().toLocaleDateString()}
-- **Project** ${data.projectTitle}
-- **Owner** ${data.projectOwner}
-
-Project risk table
--------------
-| Risk Name | Owner | Severity | Likelihood | Mitigation Status	| Risk Level | Target Date | 
-|----|----|----|----|----|----|----|
-${rows}
-`
-  return projectRiskMD;
-}
 
 /**
  * Generates a markdown report for vendor risks
